+++ CHANGELOG.md
@@ -57,7 +57,7 @@
         * [x] (DONE: 0.1.14-231217piu_a) invoice header - issue date
         * [x] (DONE: 0.1.13) invoice header - currency
         * [ ] invoice header - supplier (`<cac:AccountingSupplierParty>`)
-        * [ ] invoice header - customer (`<cac:AccountingCustomerParty>`)
+        * [ ] ...wip... invoice header - customer (`<cac:AccountingCustomerParty>`)
 
 
 
@@ -70,9 +70,11 @@
 
 * tbd... invoice customer search for other keys: "reg com", "bank / IBAN / cont", ... (area saved in `_area_to_search`)
 
-* tbd... invoice customer search and persist for "RegistrationName"
+* tbd... clean code `rdinv()` from customer area identification, line 204
 
-* wip... (last item used `240105piu_c`)
+* tbd..  next key to find: `cac:PostalAddress --> cac:Country`
+
+* 240106piu_a invoice customer search and persist for "RegistrationName"